+++ lib/lpc/buffer.h
@@ -18,6 +18,6 @@ extern buffer_t null_buf;
 
 buffer_t *null_buffer(void);
 void free_buffer(buffer_t *);
-buffer_t *allocate_buffer(int);
+buffer_t *allocate_buffer(size_t);
 int write_buffer(buffer_t *, int, char *, int);
 char *read_buffer(buffer_t *, int, int, int *);
