+++ lib/lpc/mapping.c
@@ -200,12 +200,12 @@ free_node (mapping_node_t * mn)
  *  @param n An estimate of the number of elements the mapping will hold.
  *  @return A pointer to the newly allocated mapping.
  */
-mapping_t *allocate_mapping (int n) {
+mapping_t *allocate_mapping (size_t n) {
 
   mapping_t *newmap;
   mapping_node_t **a;
 
-  if (n > CONFIG_INT (__MAX_MAPPING_SIZE__))
+  if (n > (size_t)CONFIG_INT (__MAX_MAPPING_SIZE__))
     n = CONFIG_INT (__MAX_MAPPING_SIZE__);
   newmap = ALLOCATE (mapping_t, TAG_MAPPING, "allocate_mapping: 1");
   if (newmap == NULL)
