+++ lib/lpc/mapping.h
@@ -65,7 +65,7 @@ int msameval(svalue_t *, svalue_t *);
 int mapping_save_size(mapping_t *);
 mapping_t *mapTraverse(mapping_t *, int (*) (mapping_t *, mapping_node_t *, void *), void *);
 mapping_t *load_mapping_from_aggregate(svalue_t *, int);
-mapping_t *allocate_mapping(int);
+mapping_t *allocate_mapping(size_t);
 void free_mapping(mapping_t *);
 svalue_t *find_in_mapping(mapping_t *, svalue_t *);
 svalue_t *find_string_in_mapping(mapping_t *, char *);
